PROTOCOL CHANGE. Pass the timestamp used in MsgPlayerUpdate[Small] as a double
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@18813 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/bzfsAPI.cxx
@@ -4426,7 +4426,7 @@ void bz_ServerSidePlayerHandler::handicapUpdate(int, bz_HandicapUpdateRecord **)
 
 void bz_ServerSidePlayerHandler::playerIPUpdate(int, const char*){}
 
-void bz_ServerSidePlayerHandler::playerStateUpdate(int, bz_PlayerUpdateState *, float){}
+void bz_ServerSidePlayerHandler::playerStateUpdate(int, bz_PlayerUpdateState *, double){}
 
 void bz_ServerSidePlayerHandler::playerScoreUpdate(int, float, int, int, int){}
